% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fieldplot_s3.R
\name{as.coord.data.frame}
\alias{as.coord.data.frame}
\title{Attempt to convert a data.frame to a coord.data.frame}
\usage{
as.coord.data.frame(.data, id, x, y, ...)
}
\arguments{
\item{.data}{A data.frame containing the data.}

\item{id}{Column containing the individual identifier of each coordinate point.}

\item{x}{Column containing the x-coordinate of each coordinate point.}

\item{y}{Column containing the y-coordinate of each coordinate point.}

\item{...}{Attribute columns which should be preserved.}
}
\description{
Attempt to convert a data.frame to a coord.data.frame
}
\examples{
coords <- data.frame(id1,x1,y1,d1,species1)
names(coords) <- c("id","x","y","diameter","species")
coords \%>\% as.coord.data.frame(id = id,x = x,y = y,diameter,species)
}
