% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fieldplot_s3.R
\name{cartesian_to_polar_coordinate.coord.data.frame}
\alias{cartesian_to_polar_coordinate.coord.data.frame}
\title{Convert Cartesian coord.data.frame to Polar coordinate system.}
\usage{
\method{cartesian_to_polar_coordinate}{coord.data.frame}(data)
}
\arguments{
\item{.data}{A coord.data.frame with attribute 'coord.sys' set to 'cartesian'.}
}
\value{
A coord.data.frame with attribute 'coord.sys' set to 'polar'.
}
\description{
Convert Cartesian coord.data.frame to Polar coordinate system.
}
