% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{cartesian_to_polar_coordinate.numeric}
\alias{cartesian_to_polar_coordinate.numeric}
\title{Cartesian to Polar coordinates}
\usage{
\method{cartesian_to_polar_coordinate}{numeric}(x, y)
}
\arguments{
\item{x}{x-coordinate.}

\item{y}{y-coordinate.}

\item{angle}{Logical. Default (TRUE).}
}
\value{
A list containing two vectors 'angle' and 'distance'.
}
\description{
Cartesian to Polar coordinates
}
