% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fieldplot_s3.R
\name{coordinate_reflect.coord.data.frame}
\alias{coordinate_reflect.coord.data.frame}
\title{#' Shear a coordinate system
#' @param coord.data.frame A coord.data.frame
#' @param Qx Shear factor for X-axis.
#' @param Qy Shear factor for Y-axis.
#' @return A coord.data.frame
shear_coordinates <- function(coord.data.frame,Qx=0,Qy=0){UseMethod("shear_coordinates")}}
\usage{
\method{coordinate_reflect}{coord.data.frame}(coord.data.frame, x = TRUE, y = FALSE)
}
\arguments{
\item{coord.data.frame}{A coord.data.frame}

\item{x}{Logical. (Default: TRUE) Should the coordinates be reflected across the x-axis?}

\item{y}{Logical. (Default: FALSE) Should the coordinates be reflected across the y-axis?}
}
\value{
A coord.data.frame
}
\description{

}
\details{

}
