% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fieldplot_s3.R
\name{coordinate_rotation.coord.data.frame}
\alias{coordinate_rotation.coord.data.frame}
\title{Rotation of coord.data.frame}
\usage{
\method{coordinate_rotation}{coord.data.frame}(data, rotation)
}
\arguments{
\item{rotation}{Angle of rotation, radians. e.g. [fieldcoordinate::degree_to_radian()]}
}
\value{
A coord.data.frame
}
\description{
Rotates coordinates in a coord.data.frame around the origin.
}
