% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{coordinate_rotation.numeric}
\alias{coordinate_rotation.numeric}
\title{Rotation of coordinates around origin.}
\usage{
\method{coordinate_rotation}{numeric}(x, y, rotation = 0)
}
\arguments{
\item{x}{X coordinate}

\item{y}{Y coordinate}

\item{rotation}{Angle of rotation, in radians.}
}
\value{
A list with vectors 'X' and 'Y'.
}
\description{
Rotation of coordinates around origin.
}
