% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fieldplot_s3.R
\name{coordinate_transform1}
\alias{coordinate_transform1}
\alias{coordinate_transform}
\title{Apply the transform between two images to a coord.data.frame}
\usage{
coordinate_transform(data, target_year, source_year, radius = 10, res = 0.1)
}
\arguments{
\item{data}{Coord.data.frame with points to be transformed. Must contain vars: 'Diameter','Species', 'Year', 'x','y'.}

\item{target_year}{Target coordinate year from which to produce target image.}

\item{source_year}{Coordinate year to transform. From which to produce source image.}

\item{radius}{Default (10 m radius plots) Plot radius, to calculate coordinate indices.}

\item{res}{Default (0.1) Plot resolution, to calculate coordinate indices.}
}
\value{
A coord.data.frame
}
\description{
Apply the transform between two images to a coord.data.frame
}
