% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fieldplot_s3.R
\name{coordinate_transform2}
\alias{coordinate_transform2}
\title{A shadow image of coordinate_transform that also saves images and the transform in a filepath.}
\usage{
coordinate_transform2(
  data,
  target_year,
  source_year,
  filepath = "matched_trees",
  radius = 10,
  res = 0.1
)
}
\arguments{
\item{data}{Coord.data.frame with points to be transformed. Must contain vars: 'Diameter','Species', 'Year', 'x','y'.}

\item{target_year}{Target coordinate year from which to produce target image.}

\item{source_year}{Coordinate year to transform. From which to produce source image.}

\item{filepath}{A directory to create in which to store created images and transformation files.}

\item{radius}{Default (10 m radius plots) Plot radius, to calculate coordinate indices.}

\item{res}{Default (0.1) Plot resolution, to calculate coordinate indices.}
}
\description{
This function should be internal only.
}
\keyword{internal}
