% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{degree_to_radian}
\alias{degree_to_radian}
\title{Convert an angle in degrees to radians.}
\usage{
degree_to_radian(degree)
}
\arguments{
\item{degree}{Angle, in degrees.}
}
\value{
Value of angle measured in Radians.
}
\description{
Convert an angle in degrees to radians.
}
