% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{distance_2d}
\alias{distance_2d}
\title{Calculate 2D distance between two coordinates}
\usage{
distance_2d(x1, y1, x2, y2)
}
\arguments{
\item{x1}{X of coordinate 1}

\item{y1}{Y of coordinate 1}

\item{x2}{X of coordinate 2}

\item{y2}{Y of coordinate 2}
}
\value{
A vector of 2d distances
}
\description{
Calculate 2D distance between two coordinates
}
