% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{fieldcoordinate}
\alias{fieldcoordinate}
\title{Fieldcoordinate Package}
\description{
The field coordinate package handles coordinate set tree inventories.
}
\author{
Carl Vigren <carl.vigren@slu.se>
}
