% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fieldplot_s3.R
\name{filter_radius.coord.data.frame}
\alias{filter_radius.coord.data.frame}
\title{Filter coordinate points}
\usage{
\method{filter_radius}{coord.data.frame}(.data, x = 0, y = 0, radius = 10)
}
\arguments{
\item{x}{x coordinate of point to filter from.}

\item{y}{y coordinate of point to filter from.}

\item{radius}{Maximum radius from filter point.}
}
\description{
Filter coordinate points by maximum radius (inclusive).
}
\examples{
coords <- data.frame(id1,x1,y1,d1,species1)
names(coords) <- c("id","x","y","diameter","species")
coords \%>\% as.coord.data.frame(id = id,x = x,y = y,diameter,species) \%>\% filter_radius()
}
