% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{filter_radius.numeric}
\alias{filter_radius.numeric}
\title{Circular filter for coordinate vectors}
\usage{
\method{filter_radius}{numeric}(x, y, radius = 10, ...)
}
\arguments{
\item{x}{Vector of x coordinates}

\item{y}{Vector of y coordinates}

\item{radius}{Maximum radius, inclusive.}
}
\value{
list of two vectors: 'x' & 'y'.
}
\description{
Circular filter for coordinate vectors
}
