% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fieldplot_s3.R
\name{match_rotate.coord.data.frame}
\alias{match_rotate.coord.data.frame}
\title{Rotate to best match points between two coord.data.frames}
\usage{
\method{match_rotate}{coord.data.frame}(data, coord.df2, diameter_tolerance = 3)
}
\arguments{
\item{coord.df2}{Data frame to which to rotate to.}

\item{diameter_tolerance}{Tolerance for an acceptable point match.}

\item{.data}{A coord.data.frame which to rotate.}
}
\value{
A list, containing a coord.data.frame with the rotated values; the minimising angle and the mean distance of the result rotation and coord.df2.
}
\description{
Rotate to best match points between two coord.data.frames
}
