% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fieldplot_s3.R
\name{match_trees}
\alias{match_trees}
\title{Match trees between years.}
\usage{
match_trees(
  data,
  stand_id,
  Year = "Year",
  filepath = "matched_trees/",
  save = TRUE,
  skipahead = TRUE
)
}
\arguments{
\item{data}{A data.frame}

\item{stand_id}{A character with the variable name which identifies stands.}

\item{Year}{A character with the variable name denoting different years.}

\item{filepath}{Folder in which to store matches (creates a subfolder for each stand)}

\item{save}{TRUE (default) / FALSE. Saves the transformed coordinates.}
}
\description{
Match trees between years.
}
\examples{
tree_coords2_split <- tree_coords2 \%>\% group_by(standnumber,standnr2) \%>\% group_split(.keep = TRUE)
lapply(tree_coords2_split,FUN = match_trees,stand_id = 'standnumber',Year = 'Year')
}
