% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fieldplot_s3.R
\name{most_similar_img}
\alias{most_similar_img}
\title{Find the most similar image in a folder.}
\usage{
most_similar_img(img, img_list, path = "matched_trees/mismatches/")
}
\arguments{
\item{img}{An internal Image or array (load with [png::readPNG]) which should
be transformed to the images in img_list}

\item{img_list}{A list of files contained in path, e.g. dir(path, full.names=FALSE)}

\item{path}{Path to the folder containing images to be evaluated.}
}
\value{
A list containing i) The transformed source image, ii) The target image
which returned the highest similarity, iii) A list of the similarity scores
for all evaluated images.
}
\description{
Find the most similar image in a folder.
}
