% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fieldplot_s3.R
\name{polar_to_cartesian_coordinate.coord.data.frame}
\alias{polar_to_cartesian_coordinate.coord.data.frame}
\title{Convert a Polar coord.data.frame to Cartesian coordinate system.}
\usage{
\method{polar_to_cartesian_coordinate}{coord.data.frame}(data)
}
\arguments{
\item{.data}{A coord.data.frame with attribute 'coord.sys' set to 'polar'.}
}
\value{
A coord.data.frame with attribute 'coord.sys' set to 'cartesian'.
}
\description{
Convert a Polar coord.data.frame to Cartesian coordinate system.
}
