% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{polar_to_cartesian_coordinate.numeric}
\alias{polar_to_cartesian_coordinate.numeric}
\title{Polar to Cartesian coordinates.}
\usage{
\method{polar_to_cartesian_coordinate}{numeric}(distance, theta)
}
\arguments{
\item{distance}{Distance from origin.}

\item{theta}{Angle (in radians)}
}
\value{
A list containing two vectors 'x' and 'y'.
}
\description{
Polar to Cartesian coordinates.
}
