% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fieldplot_s3.R
\name{produce_gauss_array}
\alias{produce_gauss_array}
\title{Convert a coord.data.frame to a matrix with points represented as a gaussian process.}
\usage{
produce_gauss_array(
  data,
  point_strength,
  position_error = 0.5,
  radius = 10,
  res = 0.1
)
}
\arguments{
\item{data}{A coord.data.frame}

\item{point_strength}{An attribute to be used as the amplitude of the gaussian process. Typically 'Diameter'.}

\item{position_error}{Standard deviation of gaussian process.}

\item{radius}{Radius of plot.}

\item{res}{Resolution of output matrix. Do not recommend higher resolution (lower values) than 0.1. Significant computing time.}
}
\value{
A gauss.coord.array containing columns: x,y,value.
}
\description{
Convert a coord.data.frame to a matrix with points represented as a gaussian process.
}
