% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{radian_to_degree}
\alias{radian_to_degree}
\title{Convert an angle in radians to degrees.}
\usage{
radian_to_degree(radian)
}
\arguments{
\item{radian}{Angle, in radians.}
}
\value{
Value of angle measured in degrees.
}
\description{
Convert an angle in radians to degrees.
}
