% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fieldplot_s3.R
\name{rotation_to_distance.coord.data.frame}
\alias{rotation_to_distance.coord.data.frame}
\title{Find rotation 0-360 which minimises the mean distance between matched trees.}
\usage{
\method{rotation_to_distance}{coord.data.frame}(data, rotation = 0)
}
\arguments{
\item{rotation}{Rotation to exert on coordinate (x,y) in .data, in radians.}
}
\value{
Mean distance between coordinates and their respective matches.
}
\description{
Find rotation 0-360 which minimises the mean distance between matched trees.
}
