% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fieldplot_s3.R
\name{save_gauss.default}
\alias{save_gauss.default}
\title{Save a gauss.coord.array as a png without any borders.}
\usage{
\method{save_gauss}{default}(x, filename)
}
\description{
Save a gauss.coord.array as a png without any borders.
}
