% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fieldplot_s3.R
\name{save_gauss}
\alias{save_gauss}
\title{Save a gauss.coord.array as a .png in the working directory.}
\usage{
save_gauss(x, filename)
}
\arguments{
\item{x}{A gauss.coord.array}

\item{filename}{A filename, including extension.}
}
\description{
Save a gauss.coord.array as a .png in the working directory.
}
