% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fieldplot_s3.R
\name{scale_coordinates.coord.data.frame}
\alias{scale_coordinates.coord.data.frame}
\title{Scale a coordinate system}
\usage{
\method{scale_coordinates}{coord.data.frame}(coord.data.frame, Sx, Sy)
}
\arguments{
\item{coord.data.frame}{A coord.data.frame}

\item{Sx}{Scaling factor for x-axis.}

\item{Sy}{Scaling factor for y-axis.}
}
\value{
A coord.data.frame
}
\description{
Scale a coordinate system
}
