% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fieldplot_s3.R
\name{translate_coordinates.coord.data.frame}
\alias{translate_coordinates.coord.data.frame}
\title{Translate a coordinate system.}
\usage{
\method{translate_coordinates}{coord.data.frame}(coord.data.frame, x, y)
}
\arguments{
\item{coord.data.frame}{A coord.data.frame.}

\item{x}{Points to be translated right.}

\item{y}{Points to be translated up.}
}
\value{
A coord.data.frame
}
\description{
Translate a coordinate system.
}
